feat: complete anonymyziation cycle
